pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trailcam-classifier"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Erik Abair", email = "erik.abair@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "torch~=2.8.0",
    "torchvision~=0.23.0",
    "scikit-learn~=1.7.1",
    "Pillow~=11.3.0",
]

[project.scripts]
trailcamclassify = "trailcam_classifier:run"
trailcamclassify-train = "trailcam_classifier:train"


[project.urls]
Documentation = "https://github.com/abaire/trailcam-classifier#readme"
Issues = "https://github.com/abaire/trailcam-classifier/issues"
Source = "https://github.com/abaire/trailcam-classifier"

[tool.hatch.version]
path = "src/trailcam_classifier/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/trailcam_classifier tests}"

[tool.coverage.run]
source_pkgs = ["trailcam_classifier", "tests"]
branch = true
parallel = true
omit = [
  "src/trailcam_classifier/__about__.py",
]

[tool.coverage.paths]
trailcam_classifier = ["src/trailcam_classifier", "*/trailcam-classifier/src/trailcam_classifier"]
tests = ["tests", "*/trailcam-classifier/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
